+++ src/main/java/org/moss/discord/listeners/log/LogParser.java
@@ -1,4 +1,4 @@
-package org.moss.discord.listeners.parser;
+package org.moss.discord.listeners.log;
 
 import org.javacord.api.entity.message.Message;
